+++ internal/resolvers/default/model.go
@@ -4,8 +4,8 @@ import (
 	"html"
 	"net/http"
 
+	"github.com/Chatterino/api/pkg/humanize"
 	"github.com/Chatterino/api/pkg/resolver"
-	"github.com/Chatterino/api/pkg/utils"
 	"github.com/PuerkitoBio/goquery"
 )
 
@@ -17,8 +17,8 @@ type tooltipData struct {
 }
 
 func (d *tooltipData) Truncate() {
-	d.Title = utils.TruncateString(d.Title, MaxTitleLength)
-	d.Description = utils.TruncateString(d.Description, MaxDescriptionLength)
+	d.Title = humanize.Title(d.Title)
+	d.Description = humanize.Description(d.Description)
 }
 
 func (d *tooltipData) Sanitize() {