+++ internal/resolvers/youtube/load.go
@@ -3,29 +3,15 @@ package youtube
 import (
 	"bytes"
 	"errors"
-	"fmt"
 	"log"
 	"net/http"
 	"net/url"
-	"strings"
 	"time"
 
 	"github.com/Chatterino/api/pkg/cache"
 	"github.com/Chatterino/api/pkg/humanize"
 	"github.com/Chatterino/api/pkg/resolver"
 )
-
-func parseDuration(dur string) string {
-	dur = strings.ToLower(dur)
-	dur = strings.Replace(dur, "pt", "", 1)
-	d, _ := time.ParseDuration(dur)
-	h := d / time.Hour
-	d -= h * time.Hour
-	m := d / time.Minute
-	d -= m * time.Minute
-	s := d / time.Second
-	return fmt.Sprintf("%02d:%02d:%02d", h, m, s)
-}
 
 func load(videoID string, r *http.Request) (interface{}, time.Duration, error) {
 	youtubeVideoParts := []string{
@@ -56,8 +42,8 @@ func load(videoID string, r *http.Request) (interface{}, time.Duration, error) {
 	data := youtubeTooltipData{
 		Title:        video.Snippet.Title,
 		ChannelTitle: video.Snippet.ChannelTitle,
-		Duration:     parseDuration(video.ContentDetails.Duration),
-		PublishDate:  humanize.Date("02 Jan 2006", video.Snippet.PublishedAt),
+		Duration:     humanize.DurationPT(video.ContentDetails.Duration),
+		PublishDate:  humanize.CreationDateRFC3339(video.Snippet.PublishedAt),
 		Views:        humanize.Number(video.Statistics.ViewCount),
 		LikeCount:    humanize.Number(video.Statistics.LikeCount),
 		DislikeCount: humanize.Number(video.Statistics.DislikeCount),