+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- Twitter resolver timestamp date format changed from `Jan 2 2006` to `02 Jan 2006` (#105)
 - Dev, Breaking: Moved main package from root directory to `/cmd/api`. This change also changes the path of the executable from `./api` to `./cmd/api/api` - make sure to reflect this change in your systemd unit. (#104, #107)
 - Dev: Replaced `gorilla/mux` with `go-chi/chi`. This change requires the URL parameters to be percent-encoded, specifically the slashes. (#99)
 - Added support for Livestreamfails clip links. (#98)